FE-7788: Default values from the previous sections were overriden
For each section (GraphView, DFGConfig, etc..) we were opening the file
and then closing it, sequentially, so the last section was overriding
the previous ones. Now, we write the values that were in the file when
we opened it (even in WriteOnly mode). The Read/Write logic is done in
getOrCreateValue<>.